Shim Phonegap.js for Browser Testing
I need to develop and test this app, but I do not want to constantly
update the GitHub repo for every minor change to test on a tablet. I can
write a simple shim of phonegap.js to get the app working in a desktop
browser, starting with the deviceready event.